Hopefully fixed merge conflicts.
